Completes the core logic needed with this refactor.
